requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.13.1
antlr4-tools==0.2
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
bardapi==0.1.23a0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.3
blis==0.7.11
browser-cookie3==0.19.1
bson==0.5.10
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cryptography==41.0.4
customhys==1.1.0
cycler==0.12.1
cymem==2.0.8
data==0.4
debugpy==1.8.0
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
enum34==1.1.10
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
funcsigs==1.0.2
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
google-pasta==0.2.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflect==7.0.0
install-jdk==1.1.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
jwt==1.3.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
latex==0.7.0
libclang==16.0.6
lxml==4.9.3
lz4==4.3.2
Markdown==3.5
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
optproblems==1.3
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pathy==0.10.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
pox==0.3.3
ppft==1.7.6.7
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
PyExecJS==1.5.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
scapy==2.5.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
shutilwhich==1.1.0
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
stqdm==0.0.5
streamlit==1.27.2
synthesizer==0.2.0
tempdir==0.7.1
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
translators==5.8.7
tweet-preprocessor==0.6.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
Unidecode==1.3.7
uri-template==1.3.0
urllib3==2.0.6
vaderSentiment==3.3.2
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.8
weasel==0.3.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wordcloud==1.9.2
wrapt==1.12.1
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.7.0/xx_ent_wiki_sm-3.7.0-py3-none-any.whl#sha256=66c227a793f8a79814d6ca1da7c0ae633172e2fb0a94737bc8bd2e517479e73c
yarl==1.9.2
zipp==3.17.0
